pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "lifehub-backend"
version = "0.0.1"
dependencies = [
    "requests",
    "fastapi",
    "sqlmodel@git+https://github.com/itsbekas/sqlmodel",
    "mariadb",
    "python-dotenv",
    "python-jose",
    "argon2-cffi",
]

[tool.setuptools.packages]
find = {}

[project.scripts]
lifehub-api = "lifehub.app.api:run"
lifehub-fetch = "lifehub.app.fetch:run"
lifehub-setup = "lifehub.app.setup:setup"
lifehub-clean = "lifehub.app.setup:clean"

[tool.ruff]
lint.extend-select = ["I"]
